属性制約をリファクタリング
